removed warnings

+++ Solution/Game/InputComponent.cpp
@@ -74,8 +74,8 @@ void InputComponent::Update(float aDeltaTime)
 		//Prism::Audio::AudioInterface::GetInstance()->PostEvent("Play_Laser");
 	}
 
-	myCursorPosition.x += static_cast<float>(clip(myInputWrapper->GetMouseDX(), -20, 20)) * aDeltaTime;
-	myCursorPosition.y += static_cast<float>(clip(myInputWrapper->GetMouseDY(), -20, 20)) * aDeltaTime;
+	myCursorPosition.x += static_cast<float>(clip(static_cast<float>(myInputWrapper->GetMouseDX()), -20.f, 20.f)) * aDeltaTime;
+	myCursorPosition.y += static_cast<float>(clip(static_cast<float>(myInputWrapper->GetMouseDY()), -20.f, 20.f)) * aDeltaTime;
 
 	myCursorPosition.x = clip(myCursorPosition.x, -1, 1);
 	myCursorPosition.y = clip(myCursorPosition.y, -1, 1);
